requirements.txt
fastapi
uvicorn
msal
requests
python-dotenv
twilio
pytz
apscheduler
